+++ firestore-multimodal-genai/CHANGELOG.md
@@ -1,6 +1,39 @@
+## Version 0.0.9
+
+- Add docs on regional support for Gemini APIs
+
+## Version 0.0.8
+
+- fix typo in documentation
+
+## Version 0.0.7
+
+- Update documentation
+
+- remove VARIABLE_FIELDS as a parameter
+
+## Version 0.0.6
+
+- Fixing a build issue
+
+## Version 0.0.5
+
+- Update documentation
+- Update extensions display name
+
+## Version 0.0.4
+
+- Make model a string param, to allow for future changes to model names.
+
+- Update documentation
+
+- Add safety threshold params
+
 ## Version 0.0.3
 
-- Allow Application Default Credentials Authorization
+- Add Vertex AI provider
+
+- Add image compression for large image files
 
 ## Version 0.0.2